+++ 35-Seminar-hee4/35-Seminar-hee4/App/SceneDelegate.swift
@@ -18,7 +18,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     ) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: windowScene)
-        let navigationController = UINavigationController(rootViewController: ViewController())
+        let navigationController = UINavigationController(rootViewController: SignUpViewController())
         self.window?.rootViewController = navigationController
         self.window?.makeKeyAndVisible()
     }
